requirements.txt
# Core dependencies
numpy>=1.24.0,<2.0.0
matplotlib>=3.7.0,<4.0.0
pandas>=2.0.0,<3.0.0
scikit-learn>=1.2.0,<2.0.0
scipy>=1.10.0,<2.0.0

# Image processing - Using Wand for all image operations
wand>=0.6.13,<1.0.0
opencv-python-headless>=4.5.0,<5.0.0
Pillow>=8.0.0  # Required for some torchvision transforms

# PyTorch ecosystem with --no-deps to prevent Pillow installation
torch>=2.0.0,<3.0.0 --no-deps
torchvision>=0.15.0,<1.0.0 --no-deps
torchaudio>=2.0.0,<3.0.0 --no-deps
pytorch-lightning>=1.12.0,<3.0.0

# Additional optimizers
torch-optimizer>=0.3.0

# For advanced learning rate scheduling
torch-lr-finder>=0.2.1

# For mixed precision training
torch-amp>=0.1.0

# Configuration and serialization
pyyaml>=6.0.2,<7.0.0
pydantic>=2.5.2,<3.0.0

# Development and testing
pytest>=7.0.0,<8.0.0
pytest-cov>=4.0.0,<5.0.0
black>=23.0.0,<24.0.0
isort>=5.12.0,<6.0.0
flake8>=6.0.0,<7.0.0
mypy>=1.0.0,<2.0.0

# Optional dependencies for visualization and UI
PySide6>=6.8.3,<7.0.0
tqdm>=4.65.0,<5.0.0
seaborn>=0.12.0,<1.0.0

# For model interpretation
captum>=0.6.0,<1.0.0
tensorboard>=2.11.0,<3.0.0

# For deployment
--find-links https://download.pytorch.org/whl/torch_stable.html
# Note: ONNX has version constraints based on Python version
onnx>=1.13.0,<2.0.0  # Check Python version compatibility
